Number attribute fix
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@6921 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Lang/src/model/actor/Actor.cpp
@@ -427,7 +427,7 @@ bool Actor::validate(ProblemList &problemList) const {
 
         if (attr->getAttributeType() == BaseTypes::NUM_TYPE()) {
             bool ok;
-            attr->getAttributePureValue().toString().toInt((&ok));
+            attr->getAttributePureValue().toString().toDouble((&ok));
             result &= ok;
             if (!ok) {
                 problemList << Problem(L10N::badArgument(attr->getAttributePureValue().toString()));
